refator : [STEP3] Menu, Memu group 같은 일을 한다고 판단하여 합침
