+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,0 +1,35 @@
+package kitchenpos.menu.application;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menu.dto.MenuGroupResponse;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroupResponse create(final MenuGroupRequest menuGroup) {
+        MenuGroup newGroup = new MenuGroup(menuGroup.getName());
+        return MenuGroupResponse.of(menuGroupRepository.save(newGroup));
+    }
+
+    @Transactional(readOnly = true)
+    public List<MenuGroupResponse> list() {
+        List<MenuGroup> list = menuGroupRepository.findAll();
+        return list.stream()
+            .map(MenuGroupResponse::of)
+            .collect(Collectors.toList());
+    }
+}
