--- src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,35 +1,0 @@
-package kitchenpos.menugroup.application;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
-import kitchenpos.menugroup.dto.MenuGroupRequest;
-import kitchenpos.menugroup.dto.MenuGroupResponse;
-
-@Service
-public class MenuGroupService {
-    private final MenuGroupRepository menuGroupRepository;
-
-    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
-        this.menuGroupRepository = menuGroupRepository;
-    }
-
-    @Transactional
-    public MenuGroupResponse create(final MenuGroupRequest menuGroup) {
-        MenuGroup newGroup = new MenuGroup(menuGroup.getName());
-        return MenuGroupResponse.of(menuGroupRepository.save(newGroup));
-    }
-
-    @Transactional(readOnly = true)
-    public List<MenuGroupResponse> list() {
-        List<MenuGroup> list = menuGroupRepository.findAll();
-        return list.stream()
-            .map(MenuGroupResponse::of)
-            .collect(Collectors.toList());
-    }
-}
